+++ workspaces/linkerd/plugins/linkerd/src/components/EdgesTable.tsx
@@ -32,6 +32,10 @@ const useStyles = makeStyles({
 
   gridItemCardContent: {
     flex: 1,
+    minHeight: 200,
+    alignItems: 'center',
+    justifyContent: 'center',
+    display: 'flex',
   },
   header: {
     padding: 16,