+++ workspaces/linkerd/plugins/linkerd/src/components/OctopusGraph/OctopusGraph.tsx
@@ -70,7 +70,8 @@ export const OctopusGraph = ({
       position: { x: 0, y: 0 },
       sourcePosition: Position.Left,
       targetPosition: Position.Right,
-      type: 'node',
+      selected: true,
+      type: 'default',
       data: {
         name: entity.metadata.name,
         isSource: true,
@@ -88,7 +89,7 @@ export const OctopusGraph = ({
       id: node.name,
       sourcePosition: Position.Right,
       targetPosition: Position.Left,
-      type: 'node',
+      type: 'input',
       data: {
         name: node.name,
         header: 'upstream',
@@ -117,7 +118,7 @@ export const OctopusGraph = ({
       },
       sourcePosition: Position.Left,
       targetPosition: Position.Right,
-      type: 'node',
+      type: 'output',
       position: { x: 0, y: 0 },
     });
     initialEdges.push({