+++ workspaces/linkerd/packages/app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # app
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [22c0ec6]
+  - @backstage-community/plugin-linkerd@0.3.1
+
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [04499cd]
+  - @backstage-community/plugin-linkerd@0.3.0
+
 ## 0.0.2
 
 ### Patch Changes