+++ workspaces/linkerd/plugins/linkerd-backend/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @backstage-community/plugin-linkerd-backend
 
+## 0.3.1
+
+### Patch Changes
+
+- 22c0ec6: Small tweaks to UI and fix error message in backend
+
+## 0.3.0
+
+### Minor Changes
+
+- 04499cd: Small fixes and bump to later version of Backstage
+
+## 0.2.2
+
+### Patch Changes
+
+- d53fa3d: Fix endpoint URL for in cluster API
+
+## 0.2.1
+
+### Patch Changes
+
+- ecb0e86: Fix issue with logging errors, and native in cluster API calls to control plane
+
 ## 0.2.0
 
 ### Minor Changes